# api/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_noah-isme_sma-adp-api_internal_models.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_service.CreateTeacherAssignmentRequest:
    properties:
      class_id:
        type: string
      subject_id:
        type: string
      term_id:
        type: string
    required:
    - class_id
    - subject_id
    - term_id
    type: object
  github_com_noah-isme_sma-adp-api_internal_service.CreateTeacherRequest:
    properties:
      email:
        type: string
      expertise:
        maxLength: 500
        type: string
      full_name:
        type: string
      nip:
        maxLength: 50
        type: string
      phone:
        maxLength: 50
        type: string
    required:
    - email
    - full_name
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_service.UpdateTeacherRequest:
    properties:
      active:
        type: boolean
      email:
        type: string
      expertise:
        maxLength: 500
        type: string
      full_name:
        type: string
      nip:
        maxLength: 50
        type: string
      phone:
        maxLength: 50
        type: string
    required:
    - email
    - full_name
    type: object
...
  github_com_noah-isme_sma-adp-api_pkg_errors.Error:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  github_com_noah-isme_sma-adp-api_pkg_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_errors.Error'
      meta:
        additionalProperties: true
        type: object
      pagination:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_models.Pagination'
    type: object
info:
  contact: {}
  description: Bootstrap server for Golang migration (Phase 0)
  title: SMA ADP API
  version: 0.1.0
paths:
...
  /teachers:
    get:
      parameters:
      - description: Search by name/email/NIP
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Sort field (full_name,email,created_at)
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List teachers
      tags:
      - Teachers
    post:
      consumes:
      - application/json
      parameters:
      - description: Teacher payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.CreateTeacherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create teacher
      tags:
      - Teachers
  /teachers/{id}:
    delete:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Deactivate teacher
      tags:
      - Teachers
    get:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get teacher detail
      tags:
      - Teachers
    put:
      consumes:
      - application/json
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      - description: Teacher payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.UpdateTeacherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update teacher
      tags:
      - Teachers
  /teachers/{id}/assignments:
    get:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List teacher assignments
      tags:
      - Teacher Assignments
    post:
      consumes:
      - application/json
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      - description: Assignment payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.CreateTeacherAssignmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create teacher assignment
      tags:
      - Teacher Assignments
  /teachers/{id}/assignments/{aid}:
    delete:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      - description: Assignment ID
        in: path
        name: aid
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete teacher assignment
      tags:
      - Teacher Assignments
...
  /teachers/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      - description: Status payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update teacher status
      tags:
      - Teachers
  /teachers/import:
    post:
      consumes:
      - text/csv
      parameters:
      - description: CSV document
        in: body
        name: csv
        required: true
        schema:
          type: string
      - description: Stable key for safe retries
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Import teachers from CSV
      tags:
      - Teachers
  /teachers/roster:
    get:
      parameters:
      - description: Search by name, email, or NIP
        in: query
        name: search
        type: string
      - description: Filter by status (active or inactive)
        in: query
        name: status
        type: string
      - description: Legacy alias for status
        in: query
        name: active
        type: boolean
      - description: Filter by subject
        in: query
        name: subjectId
        type: string
      - description: Filter by track
        in: query
        name: track
        type: string
      - description: Filter by availability
        in: query
        name: availability
        type: string
      - description: Filter by homeroom class
        in: query
        name: homeroomClassId
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: perPage
        type: integer
      - description: Sort field
        in: query
        name: sortField
        type: string
      - description: Sort order (ascend or descend)
        in: query
        name: sortOrder
        type: string
      - description: Legacy alias for sortField
        in: query
        name: sort
        type: string
      - description: Legacy alias for sortOrder
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List teacher roster
      tags:
      - Teachers
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"